Moved registration of modules to base class
